backend/pyproject.toml
[project]
name = "iie-backend"
version = "0.1.0"
description = "Infrastructure Intelligence Engine API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "pydantic>=2.10.0",
    "boto3>=1.35.0",
    "pillow>=10.4.0",
    "av>=12.0.0",
    "imageio>=2.36.0",
    "imageio-ffmpeg>=0.5.1",
    "python-multipart>=0.0.17",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
